requirements.txt
Flask>=1.1.0
requests>=2.20.0
pandas>=0.25.0
numpy>=1.16.0
yfinance>=0.1.50
python-dotenv>=0.10.0
flask-cors>=3.0.0
textblob>=0.15.0 